changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11777 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Dispatcher/ORDispatcherClient.h
@@ -22,7 +22,7 @@
 {
     unsigned long long amountBlocked;
 }
-- (void)netsocket:(NetSocket*)insocket dataAvailable:(unsigned)inAmount;
+- (void)netsocket:(NetSocket*)insocket dataAvailable:(NSUInteger)inAmount;
 - (void) setAmountBlocked:(unsigned long long)numBytes;
 - (unsigned long long)amountBlocked;
 - (float)percentBlocked;
